requirements.txt
qdrant-client
sentence-transformers
groq
python-dotenv
pydantic
tqdm
